requirements.txt
pillow
argparse
tqdm